% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{approx_nk_grid}
\alias{approx_nk_grid}
\title{Create a grid of points close to a target number}
\usage{
approx_nk_grid(nk, kmin = 3, kmax = 30)
}
\arguments{
\item{nk}{target number of points}

\item{kmin}{minimum number of points of the grid in one direction}

\item{kmax}{maximum number of points of the grid in one direction}
}
\value{
A matrix of locations
}
\description{
Given a target number of points, creates an M x N grid of points, such that |M-N| <= 1
}
