% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKS.r
\name{awg_knots}
\alias{awg_knots}
\title{Augment a set of knots with a grid}
\usage{
awg_knots(init_knots, xdim = c(0, 100), ydim = c(0, 100), kclose = 2)
}
\arguments{
\item{init_knots}{the initial set of knots, e.g., selected using TKS.}

\item{xdim}{the x-coordinate limits of the spatial domain}

\item{ydim}{the y-coordinate limits of the spatial domain}

\item{kclose}{the number of other knots to consider as the `neighborhood' of a knot. See details.}
}
\value{
A matrix containing a set of locations.
}
\description{
Takes a set of knot locations and augments it with a grid, so that `empty` areas are filled.
}
\details{
\code{awg_knots} computes the average distance to the next \code{kclose} knots, and creates a grid
of knots such that the minimum distance is approximately equal to this value. More precisely, let
U be the set of initial knots, and U[j,] be a particular knot. Then if \code{kclose} = 2 , the
pairwise distances between U[j,] and U[-j,] are computed, and the smallest two are averaged.
Suppose a vector \code{d} contains all these mean distances d[j], for j = 1, ..., nrow(init_knots).
Then \code{mean(d)} is the average distance to the next \code{kclose} knots.
}
\seealso{
\code{\link{TKS}}
}
