% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_mat.r
\name{basis_mat}
\alias{basis_mat}
\title{Construct matrix of basis functions}
\usage{
basis_mat(coords, knots, mult = 1.3, FUN = "bisq", ...)
}
\arguments{
\item{coords}{the coordinates of the data locations (dimension \eqn{n1 x p}).}

\item{knots}{the coordinates of the knots (dimension \eqn{n2 x p}).}

\item{mult}{multiplier for range of each basis function.}

\item{FUN}{the basis functions, currently bisquare (bisq) and modified bisquare (mod_bisq) are accepted.}

\item{...}{space for additional arguments}
}
\value{
A matrix of size \eqn{n1 \times n2} containing pairwise distances.
}
\description{
Constructs the matrix of basis functions from a set of locations and a set of knots.
}
\details{
Distances are computed using \code{\link{pairwise_dist}}, see documention of that function
for different methods of calculating distance.
}
\note{
In future releases, custom basis functions will be permitted.
}
