% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalized_robust.r
\name{dplus}
\alias{dplus}
\title{Calculate despersion, \eqn{D^{+}{D^+}}}
\usage{
dplus(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\value{
Numeric value
}
\description{
Calculate the value of the dispersion function, \eqn{D^{+}{D^+}}
}

