% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_map.r
\name{easy_map}
\alias{easy_map}
\title{A convenient mapping function}
\usage{
easy_map(data, knots = NULL, show.knots = TRUE, panel.plot = FALSE,
  panel1 = NULL, panel2 = NULL, cex = NULL, cex.knot = NULL,
  xlim = NULL, ylim = xlim, x.ticks = NULL, y.ticks = x.ticks,
  alpha = NULL, col.map = "ylgn", ...)
}
\arguments{
\item{data}{input dataset containing coordinates (columns 1 and 2) and response data (column three).}

\item{knots}{optional matrix of locations for knots.}

\item{show.knots}{logical, whether the knots should be plotted.}

\item{panel.plot}{logical, determines whether to produce a single figure or a paneled plot.}

\item{panel1}{optional, grouping factor for first paneling level.}

\item{panel2}{optional, grouping factor for second paneling level.}

\item{cex}{optional, the size of the plotted points.}

\item{cex.knot}{optional, the size of the knot locations.}

\item{xlim}{optional, the limits of the x-coordinate.}

\item{ylim}{optional, the limits of the y-coordinate.}

\item{x.ticks}{optional, the placement of tick marks on the the x-axis.}

\item{y.ticks}{optional, the placement of tick marks for the the y-axis.}

\item{alpha}{the alpha scaling factor for geom_point.}

\item{col.map}{the colormap for the plot, current options are "ylgn" (yellow-green) or "terrain.colors".}

\item{...}{space for additional arguments.}
}
\value{
A figure produced with ggplot.
}
\description{
Function that produces a simple \code{ggplot}.
}
\note{
This is a glorified wrapper for ggplot. It is unlikely to be useful in all circumstances, 
but often provides a reasonable map, especially for simulated examples using the simulation
tools in \pkg{RRSM}, or other datasets on an approximatly square domain.
}
