% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKS.r
\name{empirical_knots}
\alias{empirical_knots}
\title{Estimate knots for reduced rank spatial process}
\usage{
empirical_knots(coords, R, delta1, delta2 = delta1, slices, augment = FALSE,
  knots_grid = NULL, ...)
}
\arguments{
\item{coords}{the observed locations.}

\item{R}{the data to be used in estimating knots.}

\item{delta1}{the first bandwidth, used to determine the knots for a given threshold interval set, passed to \code{\link{threshold_knots}}.}

\item{delta2}{the second bandidth, used to detect if newly estimated knots are too close to existing knots.}

\item{slices}{the threshold cut-off values to define interval sets.}

\item{augment}{whether to augment the estimated set of knots with a regular grid of knots (fills in sparse areas). See Notes.}

\item{knots_grid}{(optional) the grid of knots to use if \code{augment=TRUE}.}

\item{...}{space for additional arguments.}
}
\value{
A matrix of locations.
}
\description{
For a given bandwith, estimate knots for a reduced rank spatial models using the centroids of interval sets, iterating through all thresholds.
}
\note{
The argument \code{augment} was considered in the course of development of the Threshold Knot 
Selection (TKS) algorithm, but was not pursued with much depth. The properties of TKS when 
augmenting with a grid have not been studied.
}
\seealso{
\code{\link{TKS}}
}
