% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{eye}
\alias{eye}
\title{Create an identity matrix}
\usage{
eye(a)
}
\arguments{
\item{a}{dimension of matrix to create.}
}
\value{
Itentity matrix of size \code{a}
}
\description{
Port of Matlab function \code{eye}.
}

