% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_loca.r
\name{gen_loca}
\alias{gen_loca}
\title{Generate locations over a spatial domain}
\usage{
gen_loca(method = "grid", ns = NULL, unit = 100, sphere = FALSE,
  x.num = NULL, y.num = x.num, alpha = 1, beta = 2.5, rad = 0.2, ...)
}
\arguments{
\item{method}{the pattern of points to generate, allowed values are grid, uniform, radiual, and spots.}

\item{ns}{the number of locations to generate.}

\item{unit}{the dimension of the domain, will be \code{unit} x \code{unit}}

\item{sphere}{logical, if \code{TRUE}, then the locations are projected onto a sphere.}

\item{x.num}{for \code{method="grid"}, the number of locations in the x-direction.}

\item{y.num}{for \code{method="grid"}, the number of locations in the y-direction.}

\item{alpha}{for \code{method="radial"}, the first parameter of the beta distribution.}

\item{beta}{for \code{method="radial"}, the second parameter of the beta distribution.}

\item{rad}{for \code{method="spots"}, the approximate (relative) radius of each spot (not on same scale as unit).}

\item{...}{space for additional arguments}
}
\value{
A matrix of locations.
}
\description{
Generates a pattern of locations over a multiple of the unit square. Various point patterns are permitted.
}
\examples{
set1 <- gen_loca( method="grid", x.num=4, y.num=5 )
set2 <- gen_loca( method="radial", ns=5000, alpha=1, beta=2.5 )

}
