% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKS.r
\name{merge_close_knots}
\alias{merge_close_knots}
\title{Merge close knots}
\usage{
merge_close_knots(knots, delta, ...)
}
\arguments{
\item{knots}{the observed locations (usually, knot locations)}

\item{delta}{the bandwidth to determine which locations are 'close'.}

\item{...}{space for additional arguments.}
}
\value{
A matrix containing a set of locations.
}
\description{
Merges (via centroid) locations that are within some bandwidth of each other.
}
\seealso{
\code{\link{TKS}}
}
