% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mom_bec.r
\name{mom_bec}
\alias{mom_bec}
\title{Binned Empirical Covariance Matrix}
\usage{
mom_bec(empirical = "cj", coords, Y, X, S, bin_centers = NULL, ...)
}
\arguments{
\item{empirical}{string identifying which empirical estimate to used for the empirical binned covariance matrix. See details.}

\item{coords}{the matrix of observed locations.}

\item{Y}{the observed data.}

\item{X}{the fixed-effects design matrix.}

\item{S}{the matrix of basis functions.}

\item{bin_centers}{matrix of coordinates for the bin centers for the MOM estimation.}

\item{...}{space for additional arguments to be passed to the estimation function.}
}
\value{
A matrix.
}
\description{
Computes the binned empirical covariance matrix for Method of Moments (MOM) estimation
in a reduced rank spatial model.
}
\details{
The Method of Moments (MOM, \code{method="MOM"}) estimation is a two-step procedure. First, an empirical binned
covariance matrix is computed, and then the covariance matrix is fitted (see \code{\link{mom_fit}}). Different options for the empirical
estimate can be selected using \code{empirical}. Currently there are four choices:
\itemize{
\item \code{"cj"}{The Cressie-Johanneson empirical binned covariance matrix.}
\item \code{"robust"}{A robust empirical binned estimator based on the median absolute deviation. }
\item \code{"dispersion"}{A robust empirical binned estimator based on a more efficient robust variance. }
\item \code{"wt_dispersion"}{A weighted version of the dispersion estimate. }
}
}
\note{
Alone, the output of this function is not useful, it should be used in tandem with \code{\link{mom_fit}}
}
\seealso{
\code{\link{rr_est}}
\code{\link{mom_fit}}
}

