% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom_fit.r
\name{mom_fit}
\alias{mom_fit}
\title{Fit the reduced dimension covariance matrix}
\usage{
mom_fit(fitting = "frobenius", SigM, Sbar, mcn = 1000, vlift = "none",
  ...)
}
\arguments{
\item{fitting}{string describing which fitting method to be used to fit the covariance matrix. See details.}

\item{SigM}{the binned empirical covariance matrix.}

\item{Sbar}{a similarly binned version of the matrix of basis functions.}

\item{mcn}{if positive, winsorizes eigenvalues of covariance matrix so that the condition number is at max \code{mcn}.}

\item{vlift}{if \code{TRUE}, lifts eigenvalues of robust covariance matrix ("proportional").}

\item{...}{space for additional arguments to be passed to the fitting function.}
}
\value{
A matrix.
}
\description{
Uses the binned empirical covariance matrix for Method of Moments (MOM) 
estimation to fit the reduced dimension covariance matrix of a reduced rank spatial model.
}
\details{
The Method of Moments (MOM, \code{method="MOM"}) estimation is a two-step procedure. 
First, an empirical binned covariance matrix is computed (see \code{\link{mom_bec}}), 
and then the covariance matrix is fitted. Different options for the fitting stage
can be selected using \code{fitting}. Currently there are two choices:
\itemize{
\item \code{"frobenius"}{The covariance is fitted by minimizing the Frobenius norm (REF).}
\item \code{"robust"}{The covariance is fitted by minimizing a robust norm (using quantile regression) (REF). }
}
}
\note{
The values of \code{SigM} and \code{Sbar} are intended to be the output values of \code{\link{mom_bec}}
}
\seealso{
\code{\link{rr_est}}
\code{\link{mom_bec}}
}
