% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_robust.r
\name{normij}
\alias{normij}
\title{Weight function for robust variance}
\usage{
normij(yin)
}
\arguments{
\item{yin}{a vector of numeric values.}
}
\value{
Numeric
}
\description{
Computes weights for the weighted dispersion based variance estimate.
}
