% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basis_mat.r
\name{pairwise_dist}
\alias{pairwise_dist}
\title{Calculate pairwise distances}
\usage{
pairwise_dist(coords1, coords2 = coords1, dmethod = SpatialTools::dist2,
  ...)
}
\arguments{
\item{coords1}{the coordinates of the first set of locations (dimension \eqn{n1 x p}).}

\item{coords2}{the coordinates of the second set of locations (dimension \eqn{n2 x p}).}

\item{dmethod}{the type of distance to compute, see Details.}

\item{...}{space for additional arguments}
}
\value{
A vector of matrix containing distances between the points.
}
\description{
Calculates pairwise distances between two sets of locations.
}
\details{
This function is a wrapper for existing distance functions. Argument \code{dmethod} selects from
two types of distances to be computed:
\itemize{
\item \code{Euclidean distance:}{ Computes distance using \code{SpatialTools::dist2}.
       For this method, \code{identical(dmethod, SpatialTools::dist2)} should return \code{TRUE} (default behavior).}
\item \code{Great circle distance:}{ Distance on a sphere/ellipsoid using \pkg{geosphere}. Argument
      \code{dmethod} should match one of the distance functions allowed for argument \code{fun} in \code{geosphere::distm}. }
}
}
\examples{
coords1 <- cbind( c(1,1,2) , c(1,2,1) )
coords2 <- cbind( c(3,3) , c(2,3) )
pairwise_dist( coords1, coords2 )
}

