% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{pSD}
\alias{pSD}
\title{Pairwise sums and differences}
\usage{
pSD(A, B)
}
\arguments{
\item{A}{a numeric vector.}

\item{B}{a numeric vector.}
}
\value{
A list with elements \code{S} and \code{D}
}
\description{
Computes the sums and differences between each pair of values of two vectors.
}

