% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_em.r, R/rr_em_constrained.r
\name{rr_em}
\alias{rr_em}
\alias{rr_em_constrained}
\title{Estimate the parameters of a spatial mixed effects model}
\usage{
rr_em(Y, X, S, epsilon = 1e-05, max_iter = 500, ...)

rr_em_constrained(Y, X, X2 = NULL, S, cone = NULL, epsilon = 1e-05,
  max_iter = 500, mySolver = "LS", ...)
}
\arguments{
\item{Y}{the observed data.}

\item{X}{the fixed-effects design matrix.}

\item{S}{the matrix of basis functions.}

\item{epsilon}{the stopping condition for EM algorithms.}

\item{max_iter}{for EM algorithms, the maximum number of iterations.}

\item{...}{space for additional arguments to be passed to the estimation function.}

\item{X2}{for constrained EM algorithm, optional design matrix for unconstrained fixed-effects.}

\item{cone}{for constrained EM algorithm, matrix of -1, 0, and 1's defining the order restriction.}

\item{mySolver}{the solver to be used in the order restriction (passed to \code{isotone::activeSet}.}
}
\description{
Computes estimates of parameters for spatial mixed effects models using an EM algorithm.
}
\note{
Some of the estimation functions do not estimate the fixed effects parameter (e.g., 
\code{rr_em_nobeta}). In this case, \code{bhat} in the return object is \code{NULL}. 
In the estimation functions of \pkg{RRSM}, if \code{bhat} is \code{NULL}, then a 
GLS estimate will be computed automatically.
}
\seealso{
\code{\link{rr_est}}
\code{\link{mom_bec}}
\code{\link{mom_fit}}
\code{\link{rr_universal_krige}}
}
