% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_em_nobeta.r, R/rr_est.r
\name{rr_em_nobeta}
\alias{rr_em_nobeta}
\alias{rr_est}
\title{Estimate the parameters of a spatial mixed effects model}
\usage{
rr_em_nobeta(Y, X, S, epsilon = 1e-05, max_iter = 500, ...)

rr_est(method = "EM", em_type = "default", empirical = "cj",
  fitting = "frobenius", Y, X, S, X2 = NULL, cone = NULL, coords = NULL,
  knots = NULL, ...)
}
\arguments{
\item{Y}{the observed data.}

\item{X}{the fixed-effects design matrix.}

\item{S}{the matrix of basis functions.}

\item{epsilon}{the stopping condition for EM algorithms.}

\item{max_iter}{for EM algorithms, the maximum number of iterations.}

\item{...}{space for additional arguments to be passed to the estimation function (see especially \code{rr_em_constrained}).}

\item{method}{string describing the type of parameter estimation; currently 
supported are Expectation-Maximization ("EM") and Method of Moments ("MOM").}

\item{em_type}{string describing further specification for the EM algorithm. See details.}

\item{empirical}{string identifying which empirical estimate to used for the empirical binned covariance matrix. See details.}

\item{fitting}{string identifying which fitting method to use for MOM estimation. See details.}

\item{X2}{for constrained EM algorithm, optional design matrix for unconstrained fixed-effects.}

\item{cone}{for constrained EM algorithm, matrix of -1, 0, and 1's defining the order restriction.}

\item{coords}{the matrix of observed locations.}

\item{knots}{the matrix of knot locations.}
}
\value{
A list with elements \code{bhat} (estimate of beta), \code{V}, \code{ssq}, \code{niter}, 
and \code{method}.
}
\description{
Computes estimates of parameters for spatial mixed effects models using either EM algorithm or
a Method of Moments estimation.
}
\details{
\code{method} is used to define a general type of estimation method.
For \code{method="EM"}, the supported options for \code{em_type} are:
\itemize{
\item \code{"default"}{A standard EM algorithm to estimate all model parameters.}
\item \code{"nobeta"}{Similar to the default method, but the fixed-effects parameter is treated not estimated.}
\item \code{"constrained"}{An EM algorithm in which the fixed effects parameter is subject to linear order constraints.}
}
The Method of Moments (MOM, \code{method="MOM"}) estimation is a two-step procedure. 
First, an empirical binned covariance matrix is computed (see \code{\link{mom_bec}}), 
and then the covariance matrix is fitted (see \code{\link{mom_fit}}. Different options 
for the empirical estimate can be selected using \code{empirical}. Currently the options are:
\itemize{
\item \code{"cj"}{The Cressie-Johanneson empirical binned covariance matrix.}
\item \code{"robust"}{A robust empirical binned estimator based on the median absolute deviation. }
\item \code{"dispersion"}{A robust empirical binned estimator based on a more efficient robust variance (computationally intensive). }
\item \code{"wt_dispersion"}{A weighted version of the dispersion estimate (computationally intensive). }
}
Additionally, there are multiple options for the fitting method (\code{fitting}):
\itemize{
\item \code{"frobenius"}{The covariance is fitted by minimizing the Frobenius norm.}
\item \code{"robust"}{The covariance is fitted by minimizing a robust norm (using quantile regression). }
}
}
\note{
Some of the estimation functions do not estimate the fixed effects parameter (e.g., 
\code{rr_em_nobeta}, and several of the MOM estimation functions). In this case, \code{bhat} 
in the return object is \code{NULL}. In the estimation functions of \pkg{RRSM}, if \code{bhat}
is \code{NULL}, then a GLS estimate will be computed automatically.
}
\seealso{
\code{\link{rr_em}}
\code{\link{mom_bec}}
\code{\link{mom_fit}}
\code{\link{rr_universal_krige}}
}
