% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_universal_krige.r
\name{rr_universal_krige}
\alias{rr_universal_krige}
\title{Universal kriging predictions for spatial process}
\usage{
rr_universal_krige(Y, X, S = NULL, coords, pgrid = coords, Xpred = X,
  Spred = S, V = NULL, ssq, tsq = 0, ncore = 1, ...)
}
\arguments{
\item{Y}{the observed data. Should already be on log-scale.}

\item{X}{the design matrix for the large-scale variation (fixed effects).}

\item{S}{the matrix of basis functions (design matrix for the random effect).}

\item{coords}{the matrix of observed locations.}

\item{pgrid}{the prediction grid.}

\item{Xpred}{the design matrix for the predicted locations.}

\item{Spred}{the matrix of basis functions for the predicted locations.}

\item{V}{the estimated covariance matrix.}

\item{ssq}{the estimated residual variance.}

\item{tsq}{the estimated variance due to dimension reduction.}

\item{ncore}{if \code{ncore>1}, number of cores to use with \pkg{doParallel} backend.}

\item{...}{space for additional arguments (passed to relelvent prediction function).}
}
\description{
Obtains spatial predictions at selected locations.
}
