% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rr_krige.r
\name{rr_predict}
\alias{rr_predict}
\title{Make spatial predictions}
\usage{
rr_predict(method = "universal", ...)
}
\arguments{
\item{method}{the prediction method to use. Currently supported are "universal" and "lognormal".}

\item{...}{space for additional arguments (passed to relelvent prediction function).}
}
\value{
A matrix with the coordinates of the predicted locations, the predicted value, and measures of uncertainty.
}
\description{
Obtains spatial predictions at selected locations
}

