\name{RRSM-package}
\alias{RRSM-package}
\alias{RRSM}
\docType{package}
\title{
\packageTitle{RRSM}
}
\description{
\packageDescription{RRSM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RRSM}
\packageIndices{RRSM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RRSM}

Maintainer: \packageMaintainer{RRSM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
