% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalized_robust.r
\name{sig2dpluswts}
\alias{sig2dpluswts}
\title{Weighted variance using dispersion}
\usage{
sig2dpluswts(x)
}
\arguments{
\item{x}{a vector of numeric values.}
}
\value{
Numeric value
}
\description{
Estimates the variance of a sample using the weighted dispersion function.
}

