% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{sigma12}
\alias{sigma12}
\title{Find square root of matrix}
\usage{
sigma12(A, ...)
}
\arguments{
\item{A}{dimension of matrix to create.}

\item{...}{space for additional arguments (currently unused).}
}
\value{
Identity matrix of size \code{a}
}
\description{
Computes the square root of a matrix by square-rooting the eigenvalues.
}
