% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rrsm.r
\name{sim_rrsm}
\alias{sim_rrsm}
\title{Simulate large spatial dataset}
\usage{
sim_rrsm(X, S, beta = NULL, V = NULL, ssq = 0.001, theta = NULL,
  coords = NULL, knots = NULL, dist = "mvrnorm", seed = NULL, ...)
}
\arguments{
\item{X}{the design matrix for the large-scale variation (fixed effects).}

\item{S}{the matrix of basis functions (design matrix for the random effect).}

\item{beta}{the large-scale variation vector (fixed-effects slopes).}

\item{V}{the covariance matrix of the reduced rank process.}

\item{ssq}{the residual variance.}

\item{theta}{the spatial dependence parameter, for if \code{V} needs to be calculated.}

\item{coords}{the observed locations, for if \code{S} needs to be calculated.}

\item{knots}{the knot locations, for if \code{S} needs to be calculated.}

\item{dist}{the distribution to use for simulating data. Currently only "mvrnorm" is supported.}

\item{seed}{optional means to control the RNG seed.}

\item{...}{space for additional arguments.}
}
\value{
A vector containing the simulated values.
}
\description{
Simulates a large spatial dataset according to a Spatial Random Effects model: \eqn{Y = X*beta + S*eta + epsilon}{Y = X\beta + S\eta + \epsilon}
}
\note{
Several arguments, including \code{S}, \code{V}, and \code{ssq} may be left empty, 
if locations and knots are provided to calculate them.
}
