% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TKS.r
\name{threshold_knots}
\alias{threshold_knots}
\title{Exceedence set knots}
\usage{
threshold_knots(coords, R, u, delta, ...)
}
\arguments{
\item{coords}{the observed locations.}

\item{R}{the data to be used to define knots.}

\item{u}{the current threshold.}

\item{delta}{the bandwidth to determine which locations are 'close'.}

\item{...}{space for additional arguments.}
}
\value{
A matrix containing a set of locations.
}
\description{
For a given bandwidth, determines the locations
}
\seealso{
\code{\link{TKS}}
\code{\link{empirical_knots}}
}

