% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TKS.r
\name{TKS}
\alias{TKS}
\title{Threshold knot selection}
\usage{
TKS(Y, X, coords, R, test_set = NULL, M = round(0.1 * nrow(X)),
  penalty = 0.5, BWs = NULL, num_bw = NULL, bw_int = c(0.03, 0.12),
  slices = NULL, num_slice = NULL, slice_int = c(0.98, 0.2),
  vpred = "EM", nres = 1, ...)
}
\arguments{
\item{Y}{the observed data. Should already be on log-scale.}

\item{X}{the design matrix for the large-scale variation (fixed effects).}

\item{coords}{the observed locations.}

\item{R}{the residuals to use to select knots (computed if missing).}

\item{test_set}{the indices of locations to use as the hold-out test set.}

\item{M}{the number of locations to use for a hold-out test set (ignored if \code{test_set} is non-null).}

\item{penalty}{the exponent in the penalty for the objective function of bandwidth selection. The MSE is weighted against the number of knots to this power.}

\item{BWs}{the bandwidths to test.}

\item{num_bw}{if \code{BWs} is null, this will set the number of bandwidths to test.}

\item{bw_int}{if \code{BWs} is null, the maximum distance in the coordinates is multiplied by the min and max of \code{bw_int} to set the minimum and maximum bandwidths to test.}

\item{slices}{the threshold cut-off values to define interval sets, defined as percentiles.}

\item{num_slice}{if \code{slices} is null, this will set the number of percentile slices.}

\item{slice_int}{if \code{slices} is null, the slices will be evenly spaces between the min and max of \code{slice_int}.}

\item{vpred}{method to estimate parameters for the validation stage of TKS.}

\item{nres}{number of resolutions of knots to seelct (currently only 1 or 2 resolutions can be handled).}

\item{...}{space for additional arguments.}
}
\value{
A matrix containing a set of locations.
}
\description{
Uses data to estimate the number and locations of knots for a reduced rank spatial process.
}
\seealso{
\code{\link{empirical_knots}}
\code{\link{threshold_knots}}
}

