% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{tr_var}
\alias{tr_var}
\title{Function used in lifting eigenvalues of}
\usage{
tr_var(aa, lam0, EigVal)
}
\arguments{
\item{aa}{a multiplier}

\item{lam0}{the smallest permissible eigenvalue, eigenvalues below this value will be lifted}

\item{EigVal}{the vector of observed eigenvalues}
}
\value{
The difference in the sum of eigenvalues
}
\description{
Compares the difference in the sum of eigenvalues
}
