% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{voronii}
\alias{voronii}
\title{Voronii}
\usage{
voronii(x)
}
\arguments{
\item{x}{vector of values (usually distances between locations and knots).}
}
\value{
Indicator vector of same length as \code{x} (typically denotes closest knot to given location).
}
\description{
For a vector of values, computes indicator vector for the minimum value. When multiple 
values are minimum, randomly selects one of the values to indicate.
}
