% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_robust.r
\name{wtii}
\alias{wtii}
\title{Weight function for robust variance}
\usage{
wtii(xin)
}
\arguments{
\item{xin}{a vector of numeric values.}
}
\value{
Numeric
}
\description{
Computes weights for the weighted dispersion based variance estimate.
}
