% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{zeros}
\alias{zeros}
\title{Create a null matrix}
\usage{
zeros(a, b = 1)
}
\arguments{
\item{a}{number of rows.}

\item{b}{number of columns.}
}
\value{
Null matrix with \code{a} rows and \code{b} columns.
}
\description{
Port of Matlab function \code{zeros}.
}
