% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00Package.r
\docType{package}
\name{RRSM}
\alias{RRSM}
\alias{RRSM-package}
\title{Reduced rank spatial models}
\description{
Various methods for reduced rank spatial analysis, including Empirical Bayesian and fully Bayesian methods.


Appropriate credit should be given when publishing results obtained using \pkg{RRSM}, or when
developing other programs/packages based off of this one. Use \code{citation(package="RRSM")} for
Bibtex information.
}
\details{
\tabular{ll}{
Package: \tab RRSM\cr
Type:    \tab Package\cr
Version: \tab 0.0-1\cr
Date:    \tab 2015-05-18\cr
License: \tab GLP-3  \cr
}
}
\author{
Casey M. Jelsema <\email{casey.jelsema@nih.gov}>
}

